refactor: add TransactionList and DecoratedTransaction
Adds an object that represents a TransactionList that can be used to
replace the different Transaction Display Data Correlator things. It has
a more OOPy api that allows for composability when building a list.

Also, introduces DecoratedTransaction which has some methods on it that
were previously duplicated across several classes.

Finally, it's allowed me to refactor some of the error messages that are
produced.

Following this I'll be moving DisplayDataCorrelator to support this new
behaviour